pyproject.toml
[project]
name = "ps_funks"
description = "function collection for plotting and general purpose functions"
version = "0.1.0"
authors = [
    { name = "Pascal P. Klamser", email = "klamser@physik.hu-berlin.de" },
]
dependencies = [
  'matplotlib',
  'numpy',
  'pathlib',
  'cycler',
  'seaborn',
  'pandas',
  'geopandas',
  'tabulate',
  'python-dotenv',
  'requests', 
  'scikit-learn', 
  'uuid', 
  'deep-translator', 
  'iso-language-codes', 
  'lingua-language-detector',
]

[project.optional-dependencies]
dev = [
    "pytest",
    "tox",
]

# this is meant to enable developer mode-install via pip
[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core >=3.2,<4"]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100

[tool.pylint.BASIC]
good-names = "df,y,X,to,cc,f,ax,i,a,b,F,GT_dy,mu,wk,k,t,estimate_R,GT_dx,tw,R_posterior,y_X_split"

[tool.pylint.TYPECHECK]
signature-mutators = "click.decorators.option"

[tool.pylint."MESSAGES CONTROL"]
disable = "missing-class-docstring,missing-function-docstring,missing-module-docstring,too-few-public-methods"
